renamed cadmium/modeling/model.hpp to cadmium/modeling/dynamic_model.hpp
