+++ include/cadmium/engine/pdevs_dynamic_simulator.hpp
@@ -25,7 +25,7 @@
  */
 
 #include <cadmium/modeling/message_bag.hpp>
-#include <cadmium/modeling/model.hpp>
+#include <cadmium/modeling/dynamic_model.hpp>
 #include <cadmium/modeling/dynamic_message_bag.hpp>
 #include <cadmium/engine/pdevs_dynamic_engine.hpp>
